improved library CDN security + preparation for data validation

+++ src/cdn.js
@@ -36,26 +36,32 @@ module.exports = function (pool) {
   router.post('/library/upload', validateToken, uploadLibraryImage.single('image'), async (req, res, next) => {
     let conn;
     try {
-      const refFilename = req.body.refFilename;
-
+      const { type, id } = req.body;
       const oldPath = req.file.path;
-      const newPath = path.join(path.dirname(oldPath), refFilename);
 
-      fs.renameSync(oldPath, newPath);
-
-      const [type, idWithExt] = refFilename.split('-');
-      const id = idWithExt.split('.')[0];
-
-      const table = type === 's' ? 'series' : 'books';
-      const idColumn = type === 's' ? 'series_id' : 'book_id';
+      const table = type === 'series' ? 'series' : 'books';
+      const idColumn = type === 'series' ? 'series_id' : 'book_id';
 
       conn = await pool.getConnection();
+      const validRef = await conn.query(
+        `SELECT EXISTS(SELECT 1 FROM ${table} WHERE ${idColumn} = ? AND user_id = ?) AS record_exists`,
+        [id, req.userId]
+      );
+
+      if (!validRef[0].record_exists) {
+        return res.error('Invalid reference or unauthorized.', 403);
+      }
+
+      const newFileName = type === 'series' ? `s-${id}.png` : `b-${id}.png`;
+      const newPath = path.join(path.dirname(oldPath), newFileName);
+      fs.renameSync(oldPath, newPath);
+
       await conn.query(
         `UPDATE ${table} SET img = ? WHERE ${idColumn} = ?`,
         [true, id]
       );
 
-      res.success({ msg: 'File uploaded.', filename: refFilename });
+      res.success({ msg: 'File uploaded.', filename: newFileName });
     } catch (err) {
       next(err);
     } finally {
@@ -84,9 +90,6 @@ module.exports = function (pool) {
 
   router.post('/users/pfp/upload', validateToken, uploadUserPFP.single('image'), async (req, res, next) => {
     try {
-      if (!req.file) {
-        return res.error('Please upload a file.', 400);
-      }
       res.success({ msg: 'File uploaded.', filename: req.file.originalname });
     } catch (err) {
       next(err);
